+++ src/components/features/record/detail/RecordDetail.tsx
@@ -1,20 +1,30 @@
 import { JSX } from "react";
 import { useRouter } from "next/router";
-import { useFetchRecord } from "./hooks/queries/useFetchRecord";
 import RecordDetailContent from "./recordDetailContent/RecordDetailContent";
 import ImageCarousel from "@/components/commons/imageCarousel/ImageCarousel";
 import BackButton from "@/components/commons/backButton/BackButton";
+import { useRecoilValue } from "recoil";
+import { loggedInUserState } from "@/shared/stores";
+import { useFetchRecordsOfMine } from "@/shared/hooks/record/useFetchRecordsOfMine";
+import { useFetchRecord } from "../hooks/useFetchRecord";
 
 export default function RecordDetail(): JSX.Element | null {
   const router = useRouter();
+  // const me = useRecoilValue(loggedInUserState);
+
   const recordId =
     router.isReady && typeof router.query.recordId === "string"
       ? router.query.recordId
       : undefined;
 
   const { data, loading, error } = useFetchRecord(recordId); // ✅ 항상 호출
-  const record = data?.fetchBoard;
+  // const { recordsOfMine, loading: recordsLoading } = useFetchRecordsOfMine();
 
+  const record = data?.fetchBoard;
+  // console.log("recordsOfMine", recordsOfMine);
+  // const isWriter=!!me && !recordsLoading&& recordsOfMine?.some((record) => record._id === recordId);
+
+  const isWriter = true; // TODO: 로그인한 유저의 id와 record의 작성자 id 비교해서 isWriter 값 설정하기
   if (!router.isReady) return null;
   if (!recordId) return null;
 
@@ -40,17 +50,16 @@ export default function RecordDetail(): JSX.Element | null {
             images={images ?? []}
             className="lg:sticky lg:top-15 lg:h-fit"
           />
-          {record && <RecordDetailContent record={record} isWriter={false} />}
+          {record && (
+            <RecordDetailContent record={record} isWriter={isWriter} />
+          )}
         </div>
       ) : (
-        record && (
-          <RecordDetailContent
-            record={record}
-            // TODO: 로그인한 유저의 id와 record의 작성자 id 비교해서 isWriter 값 설정하기
-            isWriter={true}
-            className="w-full max-w-3xl mx-auto px-2 py-2 lg:px-4"
-          />
-        )
+        <RecordDetailContent
+          record={record}
+          isWriter={isWriter}
+          className="w-full max-w-3xl mx-auto px-2 py-2 lg:px-4"
+        />
       )}
     </div>
   );
